requirements.txt
@@ .. @@
 flask
 flask-cors
 torch
 PyPDF2
 transformers
 scikit-learn
 numpy
 tqdm
 sentence-transformers 
 faiss-cpu
-
-
-
+hashlib
+datetime